server/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.28
pydantic==1.9.0
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
plotly==5.20.0
Werkzeug==2.2.2
requests
kaleido